enhancement(portfolio-and-acct-value): multiple fixes as requested
