+++ src/interfaces/hive-market.interface.ts
@@ -1,0 +1,4 @@
+export interface HiveInternalMarketLockedInOrders {
+  hive: number;
+  hbd: number;
+}
