+++ src/reference-data/account-value-type.enum.ts
@@ -1,0 +1,5 @@
+export enum AccountValueType {
+  DOLLARS = 'DOLLARS',
+  HIVE = 'HIVE',
+  HIDDEN = 'HIDDEN',
+}
